♻️ refactor: #22 터치스크롤 cityScoll컴포넌트 와 district컴포넌트 하나의 TouchScroll컴포넌트로 병합
