+++ src/components/RegionModal/TouchScroll.tsx
@@ -2,8 +2,6 @@ import React, { useEffect, useState } from 'react';
 import useRegionModel from '@/api/models/useRegion';
 import { DEFAULT_SELECTED_CITY } from '@/constants';
 import styled from 'styled-components';
-import CityScroll from './CityScroll';
-import DistrictScroll from './DistrictScroll';
 import Loading from '../Loading';
 
 const TouchScroll = () => {
@@ -15,6 +13,18 @@ const TouchScroll = () => {
     district: '',
   });
 
+  const onClickCity = (city: string) => {
+    if (region.city === city) return;
+
+    setCity(city);
+    setRegion({ city, district: '' });
+  };
+
+  const onClickDistrict = (district: string) => {
+    if (region.district === district) return;
+    setRegion((prevRegin) => ({ ...prevRegin, district }));
+  };
+
   useEffect(() => {
     getRegionData();
   }, []);
@@ -23,17 +33,28 @@ const TouchScroll = () => {
 
   return (
     <Wrapper>
-      <CityScroll
-        cities={Object.keys(regions)}
-        setCity={setCity}
-        region={region}
-        setRegion={setRegion}
-      />
-      <DistrictScroll
-        districts={regions[city]}
-        region={region}
-        setRegion={setRegion}
-      />
+      <ScrollWrapper>
+        {Object.keys(regions).map((city) => (
+          <ScrollItem
+            key={city}
+            onClick={() => onClickCity(city)}
+            isActive={city === region.city}
+          >
+            <ScrollItemTitle>{city}</ScrollItemTitle>
+          </ScrollItem>
+        ))}
+      </ScrollWrapper>
+      <ScrollWrapper>
+        {regions[city].map((district) => (
+          <ScrollItem
+            key={district}
+            onClick={() => onClickDistrict(district)}
+            isActive={region.district === district}
+          >
+            <ScrollItemTitle>{district}</ScrollItemTitle>
+          </ScrollItem>
+        ))}
+      </ScrollWrapper>
     </Wrapper>
   );
 };
@@ -48,7 +69,26 @@ const Wrapper = styled.div`
   gap: 10px;
 `;
 
-const Spinner = styled.img`
-  width: 50px;
-  height: 50px;
+const ScrollWrapper = styled.ul`
+  width: 50%;
+  height: 100%;
+  overflow-y: scroll;
+  ${({ theme }) => theme.mixins.noScrollBar}
+`;
+
+const ScrollItem = styled.li<{ isActive: boolean }>`
+  ${({ theme }) => theme.mixins.flexBox()}
+  width: 100%;
+  height: calc(100% / 3);
+  cursor: pointer;
+  &:hover {
+    background-color: ${({ theme }) => theme.color.background.lightgray};
+  }
+  background-color: ${({ isActive, theme }) =>
+    isActive && theme.color.background.lightgray};
+`;
+
+const ScrollItemTitle = styled.span`
+  font-size: 0.9rem;
+  font-weight: bold;
 `;
